Bugs fixed

+++ src/Pulses.py
@@ -21,7 +21,7 @@ def Generator(events):
   pulses['Pulse'] = 0
   pulses.Pulse = pulses.Pulse.astype(np.int8)
   pulses['Candidate'] = -1
-  pulses.Candidate = pulses.Candidate.astype(np.int16)
+  pulses.Candidate = pulses.Candidate.astype(np.int32)
   pulses['dDM'] = (gb.DM.max() - gb.DM.min()) / 2.
   pulses.dDM=pulses.dDM.astype(np.float32)
   pulses['dTime'] = (gb.Time.max() - gb.Time.min()) / 2.
